Creating an interface
In this lesson, you created an interface that created a new data type for your app. This new data type makes it possible for you to specify where HousingLocation data is required. This new data type also makes it possible for your IDE and the TypeScript compiler can ensure that HousingLocation data is used where it's required.